requirements.txt
flask==3.0.2
flask-cors==4.0.0
pyobjc-framework-Quartz==10.1
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
python-dotenv==1.0.1
werkzeug==3.0.1
click==8.1.7
blinker==1.7.0
itsdangerous==2.1.2
jinja2==3.1.3
markupsafe==2.1.5
websockets==11.0.3

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Development
black==23.7.0
flake8==6.1.0